Fri, Jun 3, 2022, 8:27 PM +00:00

+++ program2/buggy.c
@@ -10,7 +10,7 @@ int main(void)
     char c3 = '!';
 
 
-    printf("%i %i %i\n", c1, c2, c3);
+    printf("%i %i %i\n", (int) c1, (int) c2, (int) c3);
 }
